provide stable/predictable entry ordering in JSONObject
I have made a slight local customisation of the library:
use of HashMap in JSONObject has been replaced with LinkedHashMap,
so that there is a predictable sequence of object keys.

This shouldn't have any effect except making sure that when serializing
a JSONObject created from a Map, the entries are written in a
predictable order, namely the iteration order of the originating Map.
JSON objects are supposed to be unordered, but this can provide
a cosmetic improvement that's useful during debugging, diagnosis,
or generally looking at serializer output.